test: export some fixture that will be used by other modules individually
