+++ fastapi/onacut/tests/routers/test_alerts.py
@@ -1,5 +1,3 @@
-from uuid import uuid4
-
 import pytest
 
 from onacut.tests import client
@@ -17,37 +15,6 @@ def metadata():
         "lattitude": 4.012475,
         "end_time": "20:30:00",
     }
-
-
-@pytest.fixture
-def region_id() -> int:
-    # We test the post of a region directly here
-    response_region = client.post("/regions/", json={"name": str(uuid4())})
-    return response_region.json()["id"]
-
-
-@pytest.fixture
-def city_id(region_id: int) -> int:
-    # We test the post of a city directly here
-    response_city = client.post(
-        "/cities/",
-        json={
-            "name": str(uuid4()),
-            "longitude": 11.124576,
-            "lattitude": 4.012475,
-            "region_id": region_id,
-        },
-    )
-    return response_city.json()["id"]
-
-
-@pytest.fixture
-def district_id(city_id: int) -> int:
-    # We test the post of a district directly here
-    response_district = client.post(
-        "/districts/", json={"name": str(uuid4()), "city_id": city_id}
-    )
-    return response_district.json()["id"]
 
 
 def test_read_all_alerts():
